requirements.txt
Django==1.10
Pillow==4.0.0
psycopg2==2.7
django-countries==4.1
six==1.10.0
django-separatedvaluesfield==0.4.0
django-sluggable==0.4.0
django-extensions==1.7.5
PyVimeo

-e git://github.com/ulule/django-linguist.git@0bab9dd45b62f649180abe5b097dd9e67ec29039#egg=django-linguist
